refactor(web-ui): convert useShowPassword hook to TypeScript file
