--- packages/frontend/web-ui/src/common/hooks/useShowPassword.js
@@ -1,19 +1,0 @@
-import { useState } from 'react';
-
-export const useShowPassword = (initial) => {
-  const [showPassword, setShowPassword] = useState(initial);
-
-  const handleClickShowPassword = () => {
-    setShowPassword((show) => !show);
-  };
-
-  const handleMouseDownPassword = (event) => {
-    event.preventDefault();
-  };
-
-  return {
-    showPassword,
-    handleClickShowPassword,
-    handleMouseDownPassword,
-  };
-};
