+++ packages/frontend/web-ui/src/common/hooks/useShowPassword.ts
@@ -1,0 +1,21 @@
+import { useState } from 'react';
+
+export const useShowPassword = (initial: boolean) => {
+  const [showPassword, setShowPassword] = useState<boolean>(initial);
+
+  const handleClickShowPassword = () => {
+    setShowPassword((show: boolean) => !show);
+  };
+
+  const handleMouseDownPassword = (
+    event: React.FormEvent<HTMLInputElement>,
+  ) => {
+    event.preventDefault();
+  };
+
+  return {
+    showPassword,
+    handleClickShowPassword,
+    handleMouseDownPassword,
+  };
+};
